+++ crates/ty_test/src/lib.rs
@@ -5,18 +5,17 @@ use colored::Colorize;
 use mdtest::matcher::{self, Failure};
 use mdtest::parser::{self, EmbeddedFileSourceMap};
 use mdtest::{
-    Failures, FileFailures, MDTEST_TEST_FILTER, MarkdownEdit, OutputFormat, TestFile, output_format,
+    Failures, FileFailures, MDTEST_TEST_FILTER, MarkdownEdit, OutputFormat, TestFile, attempt_test,
+    output_format,
 };
 use ruff_db::Db;
 use ruff_db::cancellation::CancellationTokenSource;
 use ruff_db::diagnostic::DiagnosticId;
-use ruff_db::files::{File, FileRootKind, system_path_to_file};
-use ruff_db::panic::{PanicError, catch_unwind};
+use ruff_db::files::{FileRootKind, system_path_to_file};
 use ruff_db::system::{DbWithWritableSystem as _, SystemPath, SystemPathBuf};
 use ruff_db::testing::{setup_logging, setup_logging_with_filter};
 use ruff_diagnostics::Applicability;
 use ruff_source_file::{LineIndex, OneIndexed};
-use std::backtrace::BacktraceStatus;
 use std::fmt::Write;
 use ty_module_resolver::{
     Module, SearchPath, SearchPathSettings, list_modules, resolve_module_confident,
@@ -682,87 +681,6 @@ impl std::fmt::Display for ModuleInconsistency<'_> {
             }
         }
         Ok(())
-    }
-}
-
-/// Run a function over an embedded test file, catching any panics that occur in the process.
-///
-/// If no panic occurs, the result of the function is returned as an `Ok()` variant.
-///
-/// If a panic occurs, a nicely formatted [`FileFailures`] is returned as an `Err()` variant to
-/// be formatted into a diagnostic message by callers.
-pub fn attempt_test<'a, T, F>(
-    test_fn: F,
-    test_file: &'a TestFile<'a>,
-) -> Result<T, AttemptTestError<'a>>
-where
-    F: FnOnce(File) -> T + std::panic::UnwindSafe,
-{
-    catch_unwind(|| test_fn(test_file.file)).map_err(|info| AttemptTestError { info, test_file })
-}
-
-pub struct AttemptTestError<'a> {
-    pub info: PanicError,
-    test_file: &'a TestFile<'a>,
-}
-
-impl AttemptTestError<'_> {
-    pub fn into_file_failures(
-        self,
-        db: &dyn Db,
-        action: &str,
-        clarification: Option<&str>,
-    ) -> FileFailures {
-        let info = self.info;
-
-        let mut by_line = matcher::FailuresByLine::default();
-        let mut messages = vec![];
-        match info.location {
-            Some(location) => messages.push(Failure::new(format_args!(
-                "Attempting to {action} caused a panic at {location}"
-            ))),
-            None => messages.push(Failure::new(format_args!(
-                "Attempting to {action} caused a panic at an unknown location",
-            ))),
-        }
-        if let Some(clarification) = clarification {
-            messages.push(Failure::new(clarification));
-        }
-        messages.push(Failure::new(""));
-        match info.payload.as_str() {
-            Some(message) => messages.push(Failure::new(message)),
-            // Mimic the default panic hook's rendering of the panic payload if it's
-            // not a string.
-            None => messages.push(Failure::new("Box<dyn Any>")),
-        }
-        messages.push(Failure::new(""));
-
-        if let Some(backtrace) = info.backtrace {
-            match backtrace.status() {
-                BacktraceStatus::Disabled => {
-                    let msg = "run with `RUST_BACKTRACE=1` environment variable to \
-                         a backtrace";
-                    messages.push(Failure::new(msg));
-                }
-                BacktraceStatus::Captured => {
-                    messages.extend(backtrace.to_string().split('\n').map(Failure::new));
-                }
-                _ => {}
-            }
-        }
-
-        if let Some(backtrace) = info.salsa_backtrace {
-            salsa::attach(db, || {
-                messages.extend(format!("{backtrace:#}").split('\n').map(Failure::new));
-            });
-        }
-
-        by_line.push(OneIndexed::from_zero_indexed(0), messages);
-
-        FileFailures {
-            backtick_offsets: self.test_file.to_code_block_backtick_offsets(),
-            by_line,
-        }
     }
 }
